requirements.txt
seaborn
mysql-connector-python
pymysql
openpyxl
loguru
requests
bs4
configparser
datetime
MySQL-python
requests
csv
warnings
time
